Initial commit, PoC
Provide an interface module to execute
Proof Of Work for IOTA locally in PyOTA.

Wrapper modules use the C port of the CCurl library
and the Entangled repo implementation.

Attention!!!
Before installing the modules with 'pip install -e .',
you need to build binaries from the two repo. Use the
'init.sh' script to do so. CMake and Bazel are  build
dependencies.